Attempt to reduce processor load
Multiplied the scanStep variable by a factor of 10 and also increased
the timer interval while scanning for the base address. This greatly
reduces the amount of time the processor is under load (while PJ64 was
open and SM64 wasn't loaded, it was constantly at 25% load before, now
its quick bursts of 25% load instead). However, it is still unknown
whether this will allow the base address to be found on all computers,
so further testing will be done before I merge this with master.