Validate all textbox groups
NOTE: A bug I found is that if you enter a value in a box and then
delete it, it doesn't clear the last code-side value or address that was
validated, so clicking "Read" or "Write" still does stuff. Gonna try and
fix that next. Also, I updated some of the version numbers that I forgot
to change last time, as well as the copyright date and build date.